+++ lib/rubocop/cop/style/safe_navigation.rb
@@ -120,12 +120,21 @@ module RuboCop
             corrector.remove(begin_range(node, body))
             corrector.remove(end_range(node, body))
             corrector.insert_before(method_call.loc.dot, '&')
+            handle_comments(corrector, method_call)
 
             add_safe_nav_to_all_methods_in_chain(corrector, method_call, body)
           end
         end
 
         private
+
+        def handle_comments(corrector, method_call)
+          return if processed_source.comments.empty?
+
+          comments = processed_source.comments.map(&:text).join("\n")
+          corrector.insert_before(method_call.loc.expression,
+                                  comments + "\n")
+        end
 
         def allowed_if_condition?(node)
           node.else? || node.elsif? || node.ternary?